(feat) Implement CRUD operations for prompt directories and components
