+++ backend/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 anyhow = "1.0.95"
 argon2 = { version = "0.5.3", features = ["std"] }
 # async-openai = "0.28.2"
+regex = "1"
 
 # temporary until async-openai fixes their latest release
 async-openai = { git = "https://github.com/64bit/async-openai" }
