--- docs/package.json
@@ -1,16 +1,0 @@
-{
-  "dependencies": {
-    "vue-template-compiler": "^2.7.16"
-  },
-  "name": "docs",
-  "version": "1.0.0",
-  "main": "index.js",
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": ""
-}
